+++ src/Event/mt_evtCustomEvent.c
@@ -1197,7 +1197,7 @@ u32 FUN_003998b0(int param_1)
 
   u32 uVar4;
 
-  u64 uVar5;
+  u32 uVar5;
 
   int lVar6;
 
@@ -1985,9 +1985,9 @@ u32 FUN_0039a5c0(int param_1)
 
   u32 uVar4;
 
-  u64 uVar5;
-
-  long lVar6;
+  u32 uVar5;
+ 
+  int lVar6;
 
   u32 *puVar7;
